vcl: migrate hardware font activation and metric retrieval to FontController
Continues the decoupling of OutputDevice from font management by delegating
hardware-specific synchronization and collection-based metric retrieval to
the FontController.

Changes:
- Added FontController::ActivateFontOnDevice to encapsulate the
  direct SalGraphics::SetFont call.
- Added FontController::GetFontMetricFromCollection to handle the lazy
  initialization and flat-view retrieval of font metrics from the
  PhysicalFontCollection.
- Updated OutputDevice::InitFont and GetFontMetricFromCollection to
  delegate to the controller, ensuring the controller is lazily
  initialized if needed.
- Fixed a potential null pointer dereference in OutputDevice by
  adding safety checks for mpFontController and mpGraphics.
- Updated PhysicalFontCollection and related unit tests to use
  size_t for Count() to ensure type consistency and avoid
  unnecessary casting.

This moves OutputDevice further toward being a passive consumer of
font services rather than a manager of font engine state.

+++ vcl/inc/font/PhysicalFontCollection.hxx
@@ -53,7 +53,7 @@ public:
     // fill the list with device font faces
     void                    Add( vcl::font::PhysicalFontFace* );
     SAL_DLLPRIVATE void     Clear();
-    int                     Count() const { return maPhysicalFontFamilies.size(); }
+    size_t                  Count() const { return maPhysicalFontFamilies.size(); }
 
     // find the device font family
     vcl::font::PhysicalFontFamily* FindFontFamily( std::u16string_view rFontName ) const;
